move email gen. to separate project
