+++ src/CanaryOverflow.Email/ConfirmationEmail.cshtml
@@ -1,0 +1,24 @@
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml" lang="en">
+<head>
+  <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+  <title>Account confirmation</title>
+  <link rel="stylesheet" href="ConfirmationEmail.css">
+</head>
+
+<body>
+<table border="0" cellpadding="0" cellspacing="0" height="100%" width="100%">
+  <tr>
+    <td align="center" valign="top">
+      <table border="0" cellpadding="10" cellspacing="0" width="600" class="mail__header">
+        <tr>
+          <td align="center" valign="top">
+            <h1>Hi @Model.DisplayName! Welcome to CanaryOverflow</h1>
+          </td>
+        </tr>
+      </table>
+    </td>
+  </tr>
+</table>
+</body>
+</html>
