+++ src/CanaryOverflow.Email/ConfirmationEmail.css
@@ -1,0 +1,4 @@
+﻿.mail__header {
+  background-color: #0ea5e9;
+  color: #e4e4e7;
+}
