+++ src/CanaryOverflow.Email/ConfirmationEmailViewModel.cs
@@ -1,0 +1,3 @@
+﻿namespace CanaryOverflow.Email;
+
+public record ConfirmationEmailViewModel(string DisplayName, string ConfirmationUri);
