--- src/CanaryOverflow.Service.Mvc/Email/ConfirmationEmail.cshtml
@@ -1,17 +1,0 @@
-@using System.Web
-@using CanaryOverflow.Service.Mvc.Email
-@model ConfirmationEmailViewModel
-
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml" lang="en">
-<head>
-  <meta http-equiv="Content-Type" content="text/html; charset=utf-8"/>
-  <title>@Model.Title</title>
-  <style type="text/css">
-  </style>
-</head>
-
-<body>
-<span>Please confirm your account by <a href="@HttpUtility.HtmlEncode(Model.ConfirmationUri)">clicking here</a>.</span>
-</body>
-</html>
