--- src/CanaryOverflow.Service.Mvc/Email/ConfirmationEmailViewModel.cs
@@ -1,3 +1,0 @@
-﻿namespace CanaryOverflow.Service.Mvc.Email;
-
-public record ConfirmationEmailViewModel(string Title, string ConfirmationUri);
