+++ src/CanaryOverflow.sln
@@ -18,6 +18,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CanaryOverflow.Identity", "
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Service", "Service", "{B4AE8077-5836-41EC-877C-89856816B827}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CanaryOverflow.Email", "CanaryOverflow.Email\CanaryOverflow.Email.csproj", "{D5386D02-C6B7-4C66-800F-3A86BB1A60A8}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -48,6 +50,10 @@ Global
 		{65925B72-6D21-42FD-8CC2-90ACE44CDD60}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{65925B72-6D21-42FD-8CC2-90ACE44CDD60}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{65925B72-6D21-42FD-8CC2-90ACE44CDD60}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D5386D02-C6B7-4C66-800F-3A86BB1A60A8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D5386D02-C6B7-4C66-800F-3A86BB1A60A8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D5386D02-C6B7-4C66-800F-3A86BB1A60A8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D5386D02-C6B7-4C66-800F-3A86BB1A60A8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{1C2287F8-6D6A-49E7-B052-9B37E536F358} = {D82D2414-E462-4246-9494-B1EF116CC259}
@@ -55,5 +61,7 @@ Global
 		{CB788DED-BDDF-4F9F-AF8C-77E059E772E8} = {D82D2414-E462-4246-9494-B1EF116CC259}
 		{CA9FB94F-666D-4054-AA4A-C05F76E5A26B} = {23F82C83-3ACA-4F39-BE2A-29E09B9AC7D1}
 		{EB6A0634-D273-4FDF-8BBD-0E5AD5249E24} = {B4AE8077-5836-41EC-877C-89856816B827}
+		{D5386D02-C6B7-4C66-800F-3A86BB1A60A8} = {23F82C83-3ACA-4F39-BE2A-29E09B9AC7D1}
+		{65925B72-6D21-42FD-8CC2-90ACE44CDD60} = {B4AE8077-5836-41EC-877C-89856816B827}
 	EndGlobalSection
 EndGlobal
